+++ includes/class-tmsm-aquatonic-course-booking.php
@@ -240,6 +240,7 @@ class Tmsm_Aquatonic_Course_Booking {
 			if ( ! empty( $form_add_id ) ) {
 				//$this->loader->add_filter( 'gform_entry_post_save_'.$form_add_id, $plugin_public, 'gform_entry_post_save_booking', 20, 2 );
 				//$this->loader->add_action( 'gform_after_submission_'.$form_add_id, $plugin_public, 'gform_after_submission_booking', 20, 2 );
+				$this->loader->add_filter( 'gform_pre_render_' . $form_add_id, $plugin_public, 'gform_pre_render_add', 20, 1 );
 				$this->loader->add_action( 'gform_entry_created', $plugin_public, 'gform_entry_created', 20, 2 );
 				$this->loader->add_action( 'gform_notification_' . $form_add_id, $plugin_public, 'gform_notification_booking', 20, 3 );